Added code for multiple experimental approaches
This commit will be rollbacked and changed before ever reaching
the master branch. The current status of the code freezed unity.

I think I should do another approach, but the time for today is
running out. The approach anyhow follows:

Create a new instance of the ControllerRegisterManager in every scene
 How this is solved I am unsure, maybe on an empty gameobject.

This instance is only responsible for actually getting the controls
and saves the results into another external storage class. Before
checking for new controllers, or additional, this calls should
check with the external storage class if there already are any
controllers available, if so, just grab em and act as an proxy.

The storage class can be of singleton instance thingy thingy,
either via custom singleton, or the Singleton I've found.

Each scene used a new instance of A, Ax communicates with B.
If B has data, grab it, if not Ax should generate data via
input (press key to join), also save new data to B.
There should only be one instance of B, ever.